refactor: split update/tests.rs into 4 modules (950→≤300 LOC each)
Fixes #630 — preventive maintenance to keep tests.rs under the 1000-line
ceiling before the next non-trivial change to update/.

Split seams (matching issue proposal):

| File                    | LOC | Content                                    |
|-------------------------|-----|--------------------------------------------|
| mod.rs                  |  43 | Module declarations + SkipSignalEnvGuard   |
| classify_skip_reason.rs | 246 | CI env, agent binary, subprocess-safe,     |
|                         |     | classify_skip_reason precedence tests      |
| network.rs              | 183 | URL validation, parsing, cache, retry,     |
|                         |     | synthetic release tests                    |
| skip_line.rs            | 232 | SHA256, archive extraction, shell profile, |
|                         |     | skip-line literal contract test            |
| subcommand_routing.rs   | 254 | Subcommand → skip-check decision tests     |

All 57 update::tests pass (56 + 1 pre-existing windows-only failure
on Linux — release_target_table_covers_windows_x86_64).

Pure structural refactor — zero behavior changes.
